+++ src/Products/CPUtils/Extensions/utils.py
@@ -84,15 +84,6 @@ def get_all_site_objects(self):
     return allSiteObj
 
 
-def safe_encode(value, encoding="utf-8"):
-    """
-        Converts a value to encoding, only when it is not already encoded.
-    """
-    if isinstance(value, unicode):
-        return value.encode(encoding)
-    return value
-
-
 def sendmail(self, mfrom="", to="", body="", subject="", cc="", bcc=""):
     """
         send a mail